+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.0.1 (2023-05-18)
 
+- add a row to a an error table if there's a problem during waitlist signup
 - add support to receive an email when someone has signed up for waitlist
 
 ## v1.0.0 (2023-05-11)